Renamed `header-transforms` files and exports
